==> notebooks/02_data_cleaning.ipynb <==
import pandas as pd
import numpy as np
import re
import string
import os
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer


import nltk

nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('omw-1.4')


df = pd.read_csv("../data/processed/combined_raw_data.csv")
df.head()


stop_words = set(stopwords.words("english"))
lemmatizer = WordNetLemmatizer()

def clean_text(text):
    text = str(text).lower()
    text = re.sub(r"http\S+", "", text)
    text = re.sub(r"@\w+", "", text)
    text = re.sub(r"#\w+", "", text)
    text = re.sub(r"\d+", "", text)
    text = text.translate(str.maketrans("", "", string.punctuation))
    tokens = text.split()
    tokens = [lemmatizer.lemmatize(word) for word in tokens if word not in stop_words]
    return " ".join(tokens)


df["clean_text"] = df["text"].apply(clean_text)


before = df.shape[0]
df = df.drop_duplicates(subset="clean_text")
after = df.shape[0]

print("Removed:", before - after)


df.isnull().sum()
df = df.dropna(subset=["clean_text", "label"])


cleaned_path = "../data/processed/cleaned_data.csv"
df.to_csv(cleaned_path, index=False)
print("Saved:", cleaned_path)


df["label"].value_counts()
df.head()
