==> notebooks/06_model_training.ipynb <==
import pickle
import numpy as np
import os
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, classification_report
from sklearn.model_selection import GridSearchCV


with open("../data/processed/X_train.pkl", "rb") as f:
    X_train = pickle.load(f)

with open("../data/processed/X_test.pkl", "rb") as f:
    X_test = pickle.load(f)

with open("../data/processed/y_train.pkl", "rb") as f:
    y_train = pickle.load(f)

with open("../data/processed/y_test.pkl", "rb") as f:
    y_test = pickle.load(f)

print("Train shape:", X_train.shape)
print("Test shape:", X_test.shape)


log_reg = LogisticRegression(max_iter=1000, class_weight="balanced")


param_grid = {
    "C": [0.01, 0.1, 1, 10],
    "solver": ["liblinear"]
}

grid = GridSearchCV(
    log_reg,
    param_grid,
    cv=5,
    scoring="accuracy",
    n_jobs=-1
)

grid.fit(X_train, y_train)

print("Best Parameters:", grid.best_params_)


best_model = grid.best_estimator_
best_model


y_train_pred = best_model.predict(X_train)
train_acc = accuracy_score(y_train, y_train_pred)

print("Training Accuracy:", train_acc)


y_test_pred = best_model.predict(X_test)
test_acc = accuracy_score(y_test, y_test_pred)

print("Test Accuracy:", test_acc)


print(classification_report(y_test, y_test_pred))


os.makedirs("../artifacts/models", exist_ok=True)

model_path = "../artifacts/models/logistic_model.pkl"

with open(model_path, "wb") as f:
    pickle.dump(best_model, f)

print("Model saved at:", model_path)


print("Model Training Completed Successfully")
